Clean the bilinear/__init__.py file to private sub-modules
